+++ model.h
@@ -13,7 +13,7 @@ public:
     bool load_weights(const std::string &filename);
     bool save_weights(const std::string &filename) const;
 
-    void train(const std::vector<int> &data, int epochs, double lr, int batch_size = 1);
+    void train(const std::vector<int> &data, int epochs, float lr, int batch_size = 1);
     std::string generate(const std::vector<int> &context, int length, double temperature, bool deterministic) const;
 
 private:
@@ -24,38 +24,38 @@ private:
     int T;
     int FF;
 
-    std::vector<double> token_emb; // vocab x D
-    std::vector<double> pos_emb;   // T x D
+    std::vector<float> token_emb; // vocab x D
+    std::vector<float> pos_emb;   // T x D
 
     // Single-head self-attention block parameters.
-    std::vector<double> Wq, Wk, Wv, Wo; // D x D
+    std::vector<float> Wq, Wk, Wv, Wo; // D x D
 
     // Feed-forward network parameters.
-    std::vector<double> Wff1;              // FF x D
-    std::vector<double> bff1;              // FF
-    std::vector<double> Wff2;              // D x FF
-    std::vector<double> bff2;              // D
+    std::vector<float> Wff1;              // FF x D
+    std::vector<float> bff1;              // FF
+    std::vector<float> Wff2;              // D x FF
+    std::vector<float> bff2;              // D
 
     // LayerNorm parameters (post-attention and post-FFN).
-    std::vector<double> ln1_gamma, ln1_beta; // D
-    std::vector<double> ln2_gamma, ln2_beta; // D
+    std::vector<float> ln1_gamma, ln1_beta; // D
+    std::vector<float> ln2_gamma, ln2_beta; // D
 
     // LM head.
-    std::vector<double> Wout;              // vocab x D
-    std::vector<double> bout;              // vocab
+    std::vector<float> Wout;              // vocab x D
+    std::vector<float> bout;              // vocab
 
-    static double rand_weight();
-    static std::vector<double> softmax(const std::vector<double> &z);
-    static int sample_next_token(const std::vector<double> &logits, double temperature, bool deterministic);
-    static std::vector<double> layer_norm_forward(const std::vector<double> &x, const std::vector<double> &gamma, const std::vector<double> &beta);
-    static std::vector<double> layer_norm_backward(
-        const std::vector<double> &x,
-        const std::vector<double> &gamma,
-        const std::vector<double> &dout,
-        std::vector<double> &dgamma,
-        std::vector<double> &dbeta);
+    static float rand_weight();
+    static std::vector<float> softmax(const std::vector<float> &z);
+    static int sample_next_token(const std::vector<float> &logits, double temperature, bool deterministic);
+    static std::vector<float> layer_norm_forward(const std::vector<float> &x, const std::vector<float> &gamma, const std::vector<float> &beta);
+    static std::vector<float> layer_norm_backward(
+        const std::vector<float> &x,
+        const std::vector<float> &gamma,
+        const std::vector<float> &dout,
+        std::vector<float> &dgamma,
+        std::vector<float> &dbeta);
 
-    std::vector<double> forward_last_hidden(const std::vector<int> &tokens) const;
+    std::vector<float> forward_last_hidden(const std::vector<int> &tokens) const;
     std::vector<int> normalize_context(const std::vector<int> &tokens) const;
 };
